# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BillingCheckoutInputCurrency = typeof BillingCheckoutInputCurrency[keyof typeof BillingCheckoutInputCurrency];


export const BillingCheckoutInputCurrency = {
  usd: 'usd',
  thb: 'thb',
} as const;

export type BillingCheckoutInputPaymentMethod = typeof BillingCheckoutInputPaymentMethod[keyof typeof BillingCheckoutInputPaymentMethod];


export const BillingCheckoutInputPaymentMethod = {
  card: 'card',
  promptpay: 'promptpay',
} as const;

export interface BillingCheckoutInput {
  /**
     * Payer email sent to Stripe Checkout; credits remain assigned to the authenticated account.
     * @maxLength 320
     */
  email: string;
  /**
     * @minimum 10
     * @maximum 180
     */
  durationSeconds: number;
  currency: BillingCheckoutInputCurrency;
  paymentMethod?: BillingCheckoutInputPaymentMethod;
}

export interface BillingCheckoutResponse {
  url: string;
}

export interface BillingAccount {
  creditsRemaining: number;
  generationsUsed: number;
}

export type BillingPlansMode = typeof BillingPlansMode[keyof typeof BillingPlansMode];


export const BillingPlansMode = {
  live_pay_as_you_go: 'live_pay_as_you_go',
} as const;

export type BillingPlansRates = {
  usdPerSecond: number;
  thbPerSecond: number;
};

export type BillingPlansExamplesItem = {
  seconds: number;
  usdCents: number;
  thbSatang: number;
};

export interface BillingPlans {
  mode: BillingPlansMode;
  rates: BillingPlansRates;
  examples: BillingPlansExamplesItem[];
}

export type VideoModelCreditCosts = {[key: string]: number};

/**
 * Total credits required for videos with lip-sync, keyed by duration in seconds.
 */
export type VideoModelLipSyncCreditCosts = {[key: string]: number};

export interface VideoModel {
  id: string;
  name: string;
  provider: string;
  description: string;
  creditCosts: VideoModelCreditCosts;
  /** Total credits required for videos with lip-sync, keyed by duration in seconds. */
  lipSyncCreditCosts: VideoModelLipSyncCreditCosts;
  duration: string;
  image: string;
}

export type PredictionInputAspectRatio = typeof PredictionInputAspectRatio[keyof typeof PredictionInputAspectRatio];


export const PredictionInputAspectRatio = {
  '16:9': '16:9',
  '9:16': '9:16',
  '1:1': '1:1',
} as const;

export type PredictionInputDuration = typeof PredictionInputDuration[keyof typeof PredictionInputDuration];


export const PredictionInputDuration = {
  NUMBER_5: 5,
  NUMBER_6: 6,
  NUMBER_10: 10,
} as const;

/**
 * @nullable
 */
export type PredictionInputImagePurpose = typeof PredictionInputImagePurpose[keyof typeof PredictionInputImagePurpose] | null;


export const PredictionInputImagePurpose = {
  product: 'product',
  avatar: 'avatar',
} as const;

export type PredictionInputVoiceMode = typeof PredictionInputVoiceMode[keyof typeof PredictionInputVoiceMode];


export const PredictionInputVoiceMode = {
  none: 'none',
  text: 'text',
  upload: 'upload',
} as const;

/**
 * Voice style for AI-generated narration; ignored for uploaded recordings or no voice.
 */
export type PredictionInputVoiceStyle = typeof PredictionInputVoiceStyle[keyof typeof PredictionInputVoiceStyle];


export const PredictionInputVoiceStyle = {
  male: 'male',
  female: 'female',
  child: 'child',
} as const;

export interface PredictionInput {
  /** @minLength 3 */
  prompt: string;
  modelId: string;
  aspectRatio: PredictionInputAspectRatio;
  duration: PredictionInputDuration;
  /**
     * Optional maximum credit cost approved by the customer. If the current cost differs, do not start a prediction.
     * @minimum 1
     */
  expectedCreditCost?: number;
  /** @nullable */
  negativePrompt?: string | null;
  /** @nullable */
  imageUploadId?: string | null;
  /** @nullable */
  imagePurpose?: PredictionInputImagePurpose;
  /** Apply lip-sync to an uploaded person/avatar image with narration. */
  lipSyncEnabled?: boolean;
  voiceMode?: PredictionInputVoiceMode;
  /** Voice style for AI-generated narration; ignored for uploaded recordings or no voice. */
  voiceStyle?: PredictionInputVoiceStyle;
  /**
     * @maxLength 1200
     * @nullable
     */
  voiceText?: string | null;
  /** @nullable */
  voiceUploadId?: string | null;
}

export type SoundPromptInputKind = typeof SoundPromptInputKind[keyof typeof SoundPromptInputKind];


export const SoundPromptInputKind = {
  music: 'music',
  effects: 'effects',
} as const;

export interface SoundPromptInput {
  /**
     * @minLength 3
     * @maxLength 1000
     */
  prompt: string;
  kind: SoundPromptInputKind;
}

export type ImageUploadInputContentType = typeof ImageUploadInputContentType[keyof typeof ImageUploadInputContentType];


export const ImageUploadInputContentType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
} as const;

export interface ImageUploadInput {
  contentType: ImageUploadInputContentType;
  /**
     * @minimum 1
     * @maximum 10485760
     */
  size: number;
}

export interface ImageUpload {
  uploadId: string;
  uploadUrl: string;
  contentType: string;
}

export type VoiceUploadInputContentType = typeof VoiceUploadInputContentType[keyof typeof VoiceUploadInputContentType];


export const VoiceUploadInputContentType = {
  'audio/mpeg': 'audio/mpeg',
  'audio/mp3': 'audio/mp3',
  'audio/wav': 'audio/wav',
  'audio/x-wav': 'audio/x-wav',
  'audio/mp4': 'audio/mp4',
  'audio/x-m4a': 'audio/x-m4a',
  'audio/aac': 'audio/aac',
  'audio/ogg': 'audio/ogg',
  'audio/webm': 'audio/webm',
  'audio/flac': 'audio/flac',
  'audio/x-flac': 'audio/x-flac',
} as const;

export interface VoiceUploadInput {
  contentType: VoiceUploadInputContentType;
  /**
     * @minimum 1
     * @maximum 20971520
     */
  size: number;
}

export interface VoiceUpload {
  uploadId: string;
  uploadUrl: string;
  contentType: string;
}

export type PredictionStatus = typeof PredictionStatus[keyof typeof PredictionStatus];


export const PredictionStatus = {
  starting: 'starting',
  processing: 'processing',
  succeeded: 'succeeded',
  failed: 'failed',
  canceled: 'canceled',
} as const;

export type PredictionVoiceMode = typeof PredictionVoiceMode[keyof typeof PredictionVoiceMode];


export const PredictionVoiceMode = {
  none: 'none',
  text: 'text',
  upload: 'upload',
} as const;

export type PredictionVoiceStyle = typeof PredictionVoiceStyle[keyof typeof PredictionVoiceStyle];


export const PredictionVoiceStyle = {
  male: 'male',
  female: 'female',
  child: 'child',
} as const;

export type PredictionVoiceStatus = typeof PredictionVoiceStatus[keyof typeof PredictionVoiceStatus];


export const PredictionVoiceStatus = {
  none: 'none',
  pending: 'pending',
  processing: 'processing',
  succeeded: 'succeeded',
  failed: 'failed',
} as const;

export interface Prediction {
  id: string;
  status: PredictionStatus;
  prompt: string;
  modelId: string;
  modelName: string;
  aspectRatio: string;
  duration: number;
  /** @nullable */
  outputUrl: string | null;
  /** @nullable */
  error: string | null;
  createdAt: string;
  progress: number;
  lipSyncEnabled: boolean;
  /** @nullable */
  thumbnail?: string | null;
  voiceMode: PredictionVoiceMode;
  voiceStyle: PredictionVoiceStyle;
  voiceStatus: PredictionVoiceStatus;
  creditsRefunded: boolean;
  /** @minimum 0 */
  refundedCredits: number;
  refundPending: boolean;
}

export interface ErrorResponse {
  error: string;
}