# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /billing/account:
    get:
      operationId: getBillingAccount
      tags: [billing]
      summary: Get the signed-in user's credit balance
      responses:
        "200":
          description: Current credit balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingAccount"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /billing/checkout:
    post:
      operationId: createBillingCheckout
      tags: [billing]
      summary: Start a one-time credit purchase
      description: Creates a Stripe Checkout Session for the selected currency and payment method.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingCheckoutInput"
      responses:
        "200":
          description: Stripe Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingCheckoutResponse"
        "400":
          description: Invalid purchase details or unsupported payment method for the currency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Signed-in account is missing a primary email address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Stripe could not create the Checkout Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /predictions/{id}/sound:
    post:
      operationId: createSoundVideo
      tags: [video]
      summary: Generate a short music or sound-effects track and download the video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SoundPromptInput"
      responses:
        "200":
          description: MP4 video with generated sound
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid prompt or video not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Prediction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Sound generation or video processing failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: Sound generation timed out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /predictions/{id}/sound-upload:
    post:
      operationId: uploadSoundVideo
      tags: [video]
      summary: Add a user-owned audio track and download the video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: MP4 video with uploaded sound
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid audio file or video not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Prediction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Audio file too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Video processing failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    BillingCheckoutInput:
      type: object
      required: [email, durationSeconds, currency]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
          description: Payer email sent to Stripe Checkout; credits remain assigned to the authenticated account.
        durationSeconds:
          type: integer
          minimum: 10
          maximum: 180
        currency:
          type: string
          enum: [usd, thb]
        paymentMethod:
          type: string
          enum: [card, promptpay]
          default: card
    BillingCheckoutResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
    BillingAccount:
      type: object
      required: [creditsRemaining, generationsUsed]
      properties:
        creditsRemaining:
          type: integer
        generationsUsed:
          type: integer
...
    SoundPromptInput:
      type: object
      required: [prompt, kind]
      properties:
        prompt:
          type: string
          minLength: 3
          maxLength: 1000
        kind:
          type: string
          enum: [music, effects]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string